[bitwarden] Fix freeze when triggering bitwarden (#221)
A freeze can be observed when triggering the bitwarden plugin.
This seems to occur the second time `query.add` is called.
To fix this issue the same logic as the `kill` plugin is used: Add all
items to a local array before finally adding this using `query.add`

This fixes https://github.com/albertlauncher/python/issues/215

Signed-off-by: Darrel Griët <dgriet@gmail.com>